Implemented Task 639: Move FieldContainer and textModels to shared folder
