+++ Catfish/Areas/Applets/UI/src/lib-components/item-template-editor/components/NotificationEditor.vue
@@ -1,6 +1,6 @@
 ﻿<script lang="ts">
     import { defineComponent, PropType } from "vue";
-    import { FieldContainer } from "../models/fieldContainer";
+    import { FieldContainer } from "../../shared/models/fieldContainer";
 
     export default defineComponent({
         name: "NotificationEditor",
