+++ Catfish/Areas/Applets/UI/src/lib-components/shared/models/textModels.ts
