+++ Catfish/Catfish.csproj
@@ -122,9 +122,9 @@
   </ItemGroup>
 
   <ItemGroup>
-    <TypeScriptCompile Include="Areas\Applets\UI\src\lib-components\item-template-editor\models\fieldContainer.ts" />
+    <TypeScriptCompile Include="Areas\Applets\UI\src\lib-components\shared\models\fieldContainer.ts" />
     <TypeScriptCompile Include="Areas\Applets\UI\src\lib-components\item-template-editor\models\itemTemplate.ts" />
-    <TypeScriptCompile Include="Areas\Applets\UI\src\lib-components\item-template-editor\models\textModels.ts" />
+    <TypeScriptCompile Include="Areas\Applets\UI\src\lib-components\shared\models\textModels.ts" />
     <TypeScriptCompile Include="wwwroot\Vue\tile-grid\src\models\client-cache.ts">
       <SubType>Code</SubType>
     </TypeScriptCompile>
